[INLONG-9850][Agent] Add a function to retrieve HttpManager in the ModuleManager class
