+++ inlong-agent/agent-installer/src/main/java/org/apache/inlong/agent/installer/ModuleManager.java
@@ -21,10 +21,13 @@ import org.apache.inlong.agent.common.AbstractDaemon;
 import org.apache.inlong.agent.installer.conf.InstallerConfiguration;
 import org.apache.inlong.agent.metrics.audit.AuditUtils;
 import org.apache.inlong.agent.utils.AgentUtils;
+import org.apache.inlong.agent.utils.HttpManager;
 import org.apache.inlong.agent.utils.ThreadUtils;
 import org.apache.inlong.common.pojo.agent.installer.ConfigResult;
 import org.apache.inlong.common.pojo.agent.installer.ModuleConfig;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,23 +35,59 @@ import java.util.List;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
+import static org.apache.inlong.agent.constant.FetcherConstants.AGENT_MANAGER_REQUEST_TIMEOUT;
+import static org.apache.inlong.agent.constant.FetcherConstants.AGENT_MANAGER_VIP_HTTP_PREFIX_PATH;
+import static org.apache.inlong.agent.constant.FetcherConstants.DEFAULT_AGENT_MANAGER_REQUEST_TIMEOUT;
+import static org.apache.inlong.agent.constant.FetcherConstants.DEFAULT_AGENT_MANAGER_VIP_HTTP_PREFIX_PATH;
+
 /**
  * Installer Manager, the bridge for job manager, task manager, db e.t.c it manages agent level operations and
  * communicates with outside system.
  */
 public class ModuleManager extends AbstractDaemon {
 
+    public static final String MANAGER_ADDR = "manager.addr";
+    public static final String MANAGER_AUTH_SECRET_ID = "manager.auth.secretId";
+    public static final String MANAGER_AUTH_SECRET_KEY = "manager.auth.secretKey";
     public static final int CONFIG_QUEUE_CAPACITY = 1;
     public static final int CORE_THREAD_SLEEP_TIME = 1000;
     private static final Logger LOGGER = LoggerFactory.getLogger(ModuleManager.class);
     private final InstallerConfiguration conf;
     private final BlockingQueue<ConfigResult> configQueue;
-
-    private String currentMd5;
+    private String currentMd5 = "";
+    private String currentStoragePath = "/tmp";
+    private ClassLoader classLoader;
+    private static final GsonBuilder gsonBuilder = new GsonBuilder().setDateFormat("yyyy-MM-dd HH:mm:ss");
+    private static final Gson GSON = gsonBuilder.create();
+    private final HttpManager httpManager;
 
     public ModuleManager() {
         conf = InstallerConfiguration.getInstallerConf();
         configQueue = new LinkedBlockingQueue<>(CONFIG_QUEUE_CAPACITY);
+        classLoader = Thread.currentThread().getContextClassLoader();
+        if (classLoader == null) {
+            classLoader = ModuleManager.class.getClassLoader();
+        }
+        if (requiredKeys(conf)) {
+            httpManager = getHttpManager(conf);
+        } else {
+            throw new RuntimeException("init module manager error, cannot find required key");
+        }
+    }
+
+    public HttpManager getHttpManager(InstallerConfiguration conf) {
+        String managerAddr = conf.get(MANAGER_ADDR);
+        String managerHttpPrefixPath = conf.get(AGENT_MANAGER_VIP_HTTP_PREFIX_PATH,
+                DEFAULT_AGENT_MANAGER_VIP_HTTP_PREFIX_PATH);
+        int timeout = conf.getInt(AGENT_MANAGER_REQUEST_TIMEOUT,
+                DEFAULT_AGENT_MANAGER_REQUEST_TIMEOUT);
+        String secretId = conf.get(MANAGER_AUTH_SECRET_ID);
+        String secretKey = conf.get(MANAGER_AUTH_SECRET_KEY);
+        return new HttpManager(managerAddr, managerHttpPrefixPath, timeout, secretId, secretKey);
+    }
+
+    private boolean requiredKeys(InstallerConfiguration conf) {
+        return conf.hasKey(MANAGER_ADDR);
     }
 
     public void submitConfig(ConfigResult config) {
